+++ src/plugins/manager.ts
