--- src/telemetry/index.ts
@@ -1,3 +1,0 @@
-export {TelemetryProvider} from './TelemetryProvider';
-export {useTelemetry} from './useTelemetry';
-export {useTelemetryValue} from './useTelemetryValue';
