+++ src/telemetry/provider.tsx
@@ -1,7 +1,7 @@
 import {FC, PropsWithChildren, useMemo} from 'react';
 
-import {TelemetryContext} from './TelemetryContext';
 import {TelemetryService, TelemetryServiceOptions} from './TelemetryService';
+import {TelemetryContext} from './context';
 
 export const TelemetryProvider: FC<PropsWithChildren<TelemetryServiceOptions>> = ({children, ...options}) => {
   const value = useMemo(
